refactor(tanstack-start): consolidate tanstack vite config
