+++ packages/tanstack-start/package.json
@@ -70,6 +70,8 @@
   "dependencies": {
     "@payloadcms/translations": "workspace:*",
     "@payloadcms/ui": "workspace:*",
+    "@vitejs/plugin-react": "^6.0.0",
+    "@vitejs/plugin-rsc": "^0.5.21",
     "qs-esm": "^7.0.2"
   },
   "devDependencies": {
@@ -77,7 +79,7 @@
     "@tanstack/react-start": "^1.168.26",
     "@types/react": "^19.0.0",
     "@types/react-dom": "^19.0.0",
-    "@vitejs/plugin-react": "^6.0.0",
+    "@vitejs/plugin-rsc": "^0.5.21",
     "payload": "workspace:*",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
@@ -86,16 +88,12 @@
   "peerDependencies": {
     "@tanstack/react-router": "^1.120.0",
     "@tanstack/react-start": "^1.168.26",
-    "@vitejs/plugin-react": "^6.0.0",
     "payload": "workspace:*",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "vite": ">=8.0.0"
   },
   "peerDependenciesMeta": {
-    "@vitejs/plugin-react": {
-      "optional": true
-    },
     "vite": {
       "optional": true
     }
