+++ packages/tanstack-start/src/vite/config/optimizeDeps.ts
@@ -1,73 +1,7 @@
 /**
- * Vite-level configuration constants used by `payloadPlugin`. Kept separate so
- * `plugin.ts` stays focused on wiring.
+ * Client dep-optimizer config used by `withPayload`. Kept separate from the
+ * SSR/RSC externalization config (`./external.ts`) so each concern stays small.
  */
-
-/**
- * Server-only packages (Node-only or only ever used by the server bundle).
- * These are the Vite equivalent of Next.js's `serverExternalPackages`.
- */
-export const ssrExternalPackages: string[] = [
-  'ajv',
-  'fast-uri',
-  'drizzle-kit',
-  'drizzle-kit/api',
-  'drizzle-orm',
-  'sharp',
-  'libsql',
-  'require-in-the-middle',
-  'json-schema-to-typescript',
-  'pino',
-  'pino-pretty',
-  'graphql',
-  'mongodb',
-  'mongoose',
-  'better-sqlite3',
-  'pg',
-  'pg-native',
-  'nodemailer',
-  'aws4',
-  'pluralize',
-  'console-table-printer',
-  '@azure/storage-blob',
-  '@aws-sdk/client-s3',
-  '@aws-sdk/s3-request-presigner',
-  '@google-cloud/storage',
-]
-
-/**
- * Payload packages whose source must be processed by Vite even on the server
- * (because they are workspace `.ts` files in dev). Server-only adapters
- * (`@payloadcms/db-*`, `@payloadcms/email-*`, `@payloadcms/next`, etc.) are
- * intentionally not included — those should stay external on the SSR side.
- */
-export const payloadNoExternalPatterns: Array<RegExp | string> = [
-  '@payloadcms/ui',
-  '@payloadcms/translations',
-  '@payloadcms/tanstack-start',
-  /^@payloadcms\/richtext-lexical/,
-  /^@payloadcms\/plugin-/,
-  /^@payloadcms\/storage-/,
-]
-
-/**
- * The subset of `payloadNoExternalPatterns` that needs to participate in the
- * RSC environment. Plugins and storage adapters must be included: they register
- * custom providers/components (e.g. `plugin-multi-tenant`'s `'use client'`
- * `TenantSelectionProvider`) that render in the RSC graph via
- * `RenderServerComponent`. If the package stays external to RSC, `plugin-rsc`
- * never transforms its `'use client'` modules into client references, so the
- * component executes server-side and crashes on the first hook
- * (`Cannot read properties of null (reading 'useState')`).
- */
-export const payloadRscNoExternalPatterns: Array<RegExp | string> = [
-  '@payloadcms/ui',
-  '@payloadcms/translations',
-  '@payloadcms/tanstack-start',
-  /^@payloadcms\/richtext-lexical/,
-  /^@payloadcms\/plugin-/,
-  /^@payloadcms\/storage-/,
-]
 
 /**
  * Packages we know contain Node-only code or top-level side effects requiring
@@ -104,6 +38,12 @@ export const optimizeDepsExcludeDefaults: string[] = [
  * pre-bundled for the client. Vite's auto-discovery doesn't reliably pick
  * these up because their parent packages are in `optimizeDeps.exclude`, so we
  * list them explicitly using the `parent > child` syntax.
+ *
+ * Each entry below fixes a specific late-discovery re-optimization: without it,
+ * Vite discovers the dep *after* its initial crawl, forcing a full dep
+ * re-optimization mid-session that 404s in-flight `.vite/deps/*` chunks and
+ * remounts the admin (dropping client state). This is why the list can't simply
+ * be deleted — it is the "complete first pass" for the admin's dep graph.
  */
 export const optimizeDepsIncludeDefaults: string[] = [
   '@payloadcms/ui > sonner',
