+++ packages/tanstack-start/src/vite/workarounds/clientModuleResolution.ts
