+++ packages/tanstack-start/src/vite/workarounds/reactDomServerInRsc.ts
