+++ packages/tanstack-start/test/smoke-plugin.mjs
@@ -1,18 +1,14 @@
-// Smoke test: verifies the refactored payloadPlugin still produces a valid
-// Vite config shape and that all sub-modules wire up correctly.
-import { payloadPlugin } from '../dist/vite/plugin.js'
+// Smoke test: verifies `withPayload` still produces a valid Vite config shape
+// and that all Payload workaround sub-modules wire up correctly. Run against the
+// built output: `node packages/tanstack-start/test/smoke-plugin.mjs`.
+//
+// The React, RSC, and TanStack Start plugins are now instantiated internally, so
+// this test no longer injects mocks for them — it validates the parts of the
+// config `withPayload` owns directly.
+import { withPayload } from '../dist/exports/vite.js'
 
-const fakePlugin = { name: 'fake' }
-const fakeTanstackStart = (opts) => ({
-  name: 'tanstack-start-mock',
-  __options: opts,
-})
-
-const factory = payloadPlugin({
+const factory = withPayload({
   payloadConfigPath: '/tmp/fake-payload.config.ts',
-  reactPlugin: { ...fakePlugin, name: 'react' },
-  rscPlugin: { ...fakePlugin, name: 'rsc' },
-  tanstackStart: fakeTanstackStart,
 })
 
 const config = factory({ command: 'serve', mode: 'development' })
@@ -28,52 +24,32 @@ expectKey(config, 'environments')
 expectKey(config, 'optimizeDeps')
 expectKey(config, 'plugins')
 expectKey(config, 'resolve')
-expectKey(config, 'server')
 expectKey(config, 'ssr')
 
-const pluginNames = config.plugins.filter(Boolean).map((p) => p.name)
+const pluginNames = config.plugins
+  .filter(Boolean)
+  .flat()
+  .map((p) => p?.name)
+  .filter(Boolean)
+
+// The Payload workaround plugins we own by name — the RSC/TanStack/React plugins
+// are third-party and may be returned as arrays, so we don't assert their names.
 for (const expected of [
   'payload:client-module-resolution',
   'payload:wrap-cjs-client',
   'payload:ssr-strip-dist-style-imports',
+  'payload:react-dom-server-in-rsc',
   'payload:dev-transforms',
-  'rsc',
-  'tanstack-start-mock',
-  'react',
 ]) {
   if (!pluginNames.includes(expected)) errors.push(`missing plugin: ${expected}`)
-}
-
-const tanstackMock = config.plugins.find((p) => p?.name === 'tanstack-start-mock')
-if (!tanstackMock?.__options?.importProtection?.client?.specifiers?.length) {
-  errors.push('importProtection.client.specifiers not wired')
-}
-if (
-  !tanstackMock?.__options?.importProtection?.server ||
-  !Array.isArray(tanstackMock.__options.importProtection.server.files) ||
-  tanstackMock.__options.importProtection.server.files.length !== 0
-) {
-  errors.push('importProtection.server.files must be [] (disables *.client.* SSR self-denial)')
-}
-if (!tanstackMock?.__options?.rsc?.enabled) {
-  errors.push('rsc.enabled not set')
-}
-if (tanstackMock?.__options?.router?.autoCodeSplitting !== false) {
-  errors.push('router.autoCodeSplitting must be false (so SSR routes hydrate eagerly)')
-}
-const splitGroupings = tanstackMock?.__options?.router?.codeSplittingOptions?.defaultBehavior
-if (!Array.isArray(splitGroupings) || splitGroupings.length !== 0) {
-  errors.push(
-    'router.codeSplittingOptions.defaultBehavior must be [] (the autoCodeSplitting flag is silently dropped by tanstackStart, so this is the actual mechanism that disables ?tsr-split=component lazy chunks)',
-  )
 }
 
 if (!config.ssr.external.includes('drizzle-kit')) errors.push('ssr.external missing drizzle-kit')
 if (!config.ssr.noExternal.some((p) => p === '@payloadcms/ui'))
   errors.push('ssr.noExternal missing @payloadcms/ui')
 
-if (!config.optimizeDeps.include.includes('scheduler')) {
-  errors.push('optimizeDeps.include missing scheduler')
+if (!config.optimizeDeps.include.includes('react-dom > scheduler')) {
+  errors.push('optimizeDeps.include missing react-dom > scheduler')
 }
 if (!config.optimizeDeps.exclude.includes('payload')) {
   errors.push("optimizeDeps.exclude missing 'payload'")
@@ -82,10 +58,35 @@ if (!config.optimizeDeps.exclude.includes('@payloadcms/ui')) {
   errors.push("optimizeDeps.exclude missing '@payloadcms/ui'")
 }
 
+// The `vite` override must be merged on top of the defaults: arrays appended,
+// objects deep-merged, and the Payload base preserved.
+const mergedFactory = withPayload({
+  payloadConfigPath: '/tmp/fake-payload.config.ts',
+  vite: {
+    optimizeDeps: { include: ['recharts'] },
+    server: { port: 4000 },
+  },
+})
+const merged = mergedFactory({ command: 'serve', mode: 'development' })
+
+if (merged.server?.port !== 4000) {
+  errors.push('vite override not merged: server.port !== 4000')
+}
+if (!merged.optimizeDeps.include.includes('recharts')) {
+  errors.push('vite override not appended: optimizeDeps.include missing recharts')
+}
+if (!merged.optimizeDeps.include.includes('react-dom > scheduler')) {
+  errors.push('vite override clobbered base: optimizeDeps.include lost defaults')
+}
+
 if (errors.length) {
   console.error('FAIL\n' + errors.map((e) => ' - ' + e).join('\n'))
   process.exit(1)
 }
-console.log('OK: payloadPlugin produces expected config shape')
-console.log(`  plugins=${pluginNames.length}  ssr.external=${config.ssr.external.length}  ssr.noExternal=${config.ssr.noExternal.length}`)
-console.log(`  optimizeDeps.include=${config.optimizeDeps.include.length}  optimizeDeps.exclude=${config.optimizeDeps.exclude.length}`)
+console.log('OK: withPayload produces expected config shape')
+console.log(
+  `  ssr.external=${config.ssr.external.length}  ssr.noExternal=${config.ssr.noExternal.length}`,
+)
+console.log(
+  `  optimizeDeps.include=${config.optimizeDeps.include.length}  optimizeDeps.exclude=${config.optimizeDeps.exclude.length}`,
+)
